+++ Integration/AwsEndUserMessagingSms/ConfigurationProviderInterface.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+namespace MauticPlugin\AwsEndUserMessagingSmsBundle\Integration\AwsEndUserMessagingSms;
+
+interface ConfigurationProviderInterface
+{
+    /**
+     * @return array<string, mixed>
+     */
+    public function get(): array;
+}